// generated/prisma/models/Asset.ts
export type AssetModel = runtime.Types.Result.DefaultSelection<Prisma.$AssetPayload>
// ...
export type $AssetPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "Asset"
  objects: {
    property: Prisma.$PropertyPayload<ExtArgs>
    shares: Prisma.$LeadAssetSharePayload<ExtArgs>[]
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    propertyId: string
    title: string
    description: string | null
    type: $Enums.AssetType
    category: $Enums.AssetCategory
    url: string
    thumbnailUrl: string | null
    tags: string[]
    isActive: boolean
    createdBy: string | null
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["asset"]>
  composites: {}
}